Establish minikube environment for spangen
This work is intended to allow other folks to see the issues I have with
istio / jaeger / opencensus-java in a reproducible minikube environment.
As of this commit minikube is... not quite working, but I'll ask about
that specifically. It's just a touch different than actual kubernetes
and that leads to odd failures you wouldn't see in an actual cluster.

Actually, while preparing the README it seems like there's some serious
flaw in the setup that might explain the lack of traces. That's the next
line of investigation.